requirements.txt
absl-py==0.12.0
APScheduler==3.6.3
build==0.10.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
cycler==0.10.0
decorator==5.1.1
et-xmlfile==1.1.0
fanstatic==1.2
idna==3.4
imageio==2.9.0
importlib-metadata==6.0.0
joblib==1.0.1
js==1.0
kiwisolver==1.3.1
log2d==0.0.18
matplotlib==3.4.2
numpy==1.20.3
openpyxl==3.0.7
packaging==21.3
pandas==1.2.4
patsy==0.5.2
Pillow==8.3.0
psutil==5.8.0
pycparser==2.21
pyparsing==2.4.7
pyproject-hooks==1.0.0
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2021.1
requests==2.28.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.12.0
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==2.1.0
tomli==2.0.1
torch==1.9.0
tornado==6.1
tqdm==4.61.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.13
WebOb==1.8.7
xlrd==2.0.1
zipp==3.12.1
